+++ quint/src/cli.ts
@@ -9,35 +9,38 @@
  * @author Igor Konnov, Gabriela Moreira, Shon Feder, Informal Systems, 2021-2023
  */
 
-
 import yargs from 'yargs/yargs'
 
 import {
-  docs, load, outputResult, parse, runRepl, runSimulator, runTests, typecheck
+  docs,
+  load,
+  outputResult,
+  parse,
+  runRepl,
+  runSimulator,
+  runTests,
+  typecheck
 } from './cliCommands'
 
 import { verbosity } from './verbosity'
 
 import { version } from './version'
 
+const defaultOpts = (yargs: any) =>
+  yargs.option('out', {
+    desc: 'output file (suppresses all console output)',
+    type: 'string',
+  })
+
 // construct parsing commands with yargs
 const parseCmd = {
   command: 'parse <input>',
   desc: 'parse a Quint specification',
   builder: (yargs: any) =>
-    yargs
-      .option('out', {
-        desc: 'output file',
-        type: 'string',
-      })
-      .option('source-map', {
-        desc: 'name of the source map',
-        type: 'string',
-      })
-      .option('with-lookup', {
-        desc: 'add the lookup table to the output file (see --out)',
-        type: 'boolean',
-      }),
+    defaultOpts(yargs).option('source-map', {
+      desc: 'name of the source map',
+      type: 'string',
+    }),
   handler: (args: any) => outputResult(load(args).chain(parse)),
 }
 
@@ -45,13 +48,9 @@ const parseCmd = {
 const typecheckCmd = {
   command: 'typecheck <input>',
   desc: 'check types and effects of a Quint specification',
-  builder: (yargs: any) =>
-    yargs
-      .option('out', {
-        desc: 'output file',
-        type: 'string',
-      }),
-  handler: (args: any) => outputResult(load(args).chain(parse).chain(typecheck)),
+  builder: defaultOpts,
+  handler: (args: any) =>
+    outputResult(load(args).chain(parse).chain(typecheck)),
 }
 
 // construct repl commands with yargs
@@ -61,12 +60,14 @@ const replCmd = {
   builder: (yargs: any) =>
     yargs
       .option('require', {
-        desc: 'filename[::module]. Preload the file and, optionally, import the module',
+        desc:
+          'filename[::module]. Preload the file and, optionally, import the module',
         alias: 'r',
         type: 'string',
       })
       .option('quiet', {
-        desc: 'Disable banners and prompts, to simplify scripting (alias for --verbosity=0)',
+        desc:
+          'Disable banners and prompts, to simplify scripting (alias for --verbosity=0)',
         alias: 'q',
         type: 'boolean',
       })
@@ -84,17 +85,14 @@ const testCmd = {
   command: 'test <input>',
   desc: 'Run tests against a Quint specification',
   builder: (yargs: any) =>
-    yargs
+    defaultOpts(yargs)
       .option('main', {
         desc: 'name of the main module (by default, computed from filename)',
         type: 'string',
       })
-      .option('out', {
-        desc: 'output file (suppresses all console output)',
-        type: 'string',
-      })
       .option('max-samples', {
-        desc: 'the maximum number of successful runs to try for every randomized test',
+        desc:
+          'the maximum number of successful runs to try for every randomized test',
         type: 'number',
       })
       .default('max-samples', 10000)
@@ -107,13 +105,13 @@ const testCmd = {
         type: 'number',
       })
       .default('verbosity', verbosity.defaultLevel)
-// Timeouts are postponed for:
-// https://github.com/informalsystems/quint/issues/633
-//
-//      .option('timeout', {
-//        desc: 'timeout in seconds',
-//        type: 'number',
-//      })
+      // Timeouts are postponed for:
+      // https://github.com/informalsystems/quint/issues/633
+      //
+      //      .option('timeout', {
+      //        desc: 'timeout in seconds',
+      //        type: 'number',
+      //      })
       .option('match', {
         desc: 'a string or regex that selects names to use as tests',
         type: 'string',
@@ -127,17 +125,14 @@ const runCmd = {
   command: 'run <input>',
   desc: 'Simulate a Quint specification and check invariants',
   builder: (yargs: any) =>
-    yargs
+    defaultOpts(yargs)
       .option('main', {
         desc: 'name of the main module (by default, computed from filename)',
         type: 'string',
       })
-      .option('out', {
-        desc: 'output file (suppresses all console output)',
-        type: 'string',
-      })
       .option('out-itf', {
-        desc: 'output the trace in the Informal Trace Format to file (supresses all console output)',
+        desc:
+          'output the trace in the Informal Trace Format to file (supresses all console output)',
         type: 'string',
       })
       .option('max-samples', {
@@ -174,13 +169,13 @@ const runCmd = {
         type: 'number',
       })
       .default('verbosity', verbosity.defaultLevel),
-// Timeouts are postponed for:
-// https://github.com/informalsystems/quint/issues/633
-//
-//      .option('timeout', {
-//        desc: 'timeout in seconds',
-//        type: 'number',
-//      })
+  // Timeouts are postponed for:
+  // https://github.com/informalsystems/quint/issues/633
+  //
+  //      .option('timeout', {
+  //        desc: 'timeout in seconds',
+  //        type: 'number',
+  //      })
   handler: (args: any) =>
     outputResult(load(args).chain(parse).chain(typecheck).chain(runSimulator)),
 }
@@ -189,14 +184,10 @@ const runCmd = {
 const docsCmd = {
   command: 'docs <input>',
   desc: 'produces documentation from docstrings in a Quint specification',
-  builder: (yargs: any) =>
-    yargs
-      .option('out', {
-        desc: 'output file',
-        type: 'string',
-      }),
+  builder: defaultOpts,
   handler: (args: any) => outputResult(load(args).chain(docs)),
 }
+
 // parse the command-line arguments and execute the handlers
 yargs(process.argv.slice(2))
   .command(parseCmd)