+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add the precursor of the standard library in the form of spells (#827)
 - cache top-level `pure val` (#837)
 - support for higher-order operators in REPL/simulator (#845)
+- More data is now output in most commands when `--out` is provided (#852)
 
 ### Changed
 
@@ -29,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `notin` is no longer a builtin operator (#814)
+- `--with-lookup` command (#850)
 
 ### Fixed