+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cache top-level `pure val` (#837)
 - support for higher-order operators in REPL/simulator (#845)
 - More data is now output in most commands when `--out` is provided (#852)
+- Mock command `verify` for running Apalache in the future (#787)
 
 ### Changed