+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add the precursor of the standard library in the form of spells (#827)
 - cache top-level `pure val` (#837)
 - support for higher-order operators in REPL/simulator (#845)
+- introduce `reps` as an alternative to `repeated` (#855)
 - More data is now output in most commands when `--out` is provided (#852)
 - Mock command `verify` for running Apalache in the future (#787)
 
@@ -28,6 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The ITF export adds output similar to Apalache (#780)
 
 ### Deprecated
+
+- `repeated` is deprecated in favor of `reps` (#855)
+
 ### Removed
 
 - `notin` is no longer a builtin operator (#814)